modify tests, move APITest to app/tests.py
